+++ R/zzz.R
@@ -1,11 +1,13 @@
 .onLoad <- function(libname, pkgname) { # nolint
   op <- options()
   op_covr <- list(
+    covr.covrignore = Sys.getenv("COVR_COVRIGNORE", ".covrignore"),
     covr.gcov = Sys.which("gcov"),
     covr.gcov_args = NULL,
     covr.exclude_pattern = rex::rex("#", any_spaces, "nocov"),
     covr.exclude_start = rex::rex("#", any_spaces, "nocov", any_spaces, "start"),
     covr.exclude_end = rex::rex("#", any_spaces, "nocov", any_spaces, "end"),
+    covr.cc = Sys.which("gcc"),
     covr.flags = c(CFLAGS = "-O0 --coverage",
                  CXXFLAGS = "-O0 --coverage",
                  CXX1XFLAGS = "-O0 --coverage",